Document planned otpions, adjust parser scripts.
Parser was adjusted to deal with --no-rom
option, which is really just a --rom option
negated. Also had to deal with a new, unusual
option-argument string in README.md.

gen-names.awk should have been named
gen-options.awk from the beginning, and it has
confused me a few times, so I moved it.